requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu118

fastapi==0.95.2
uvicorn[standard]==0.22.0
pydantic==1.10.11
python-multipart==0.0.6
cython==3.0
openai-whisper==20250625
torch==2.7.1
torchvision==0.22.1+cu118
torchaudio==2.7.1+cu118
pydub==0.25.1
soundfile==0.12.1
pytest==7.4.0
httpx==0.24.1
pytest-asyncio==0.22.0
gunicorn==20.1.0
